transaction validation, load disbursement

+++ src/main/java/ke/co/myfuture/Myfuture/Treasury/Loan/CreateLoanRequest.java
@@ -29,7 +29,7 @@ public class CreateLoanRequest {
 
     @NotNull
     @DecimalMin("0.0")
-    private BigDecimal requestedAmount;
+    private Double requestedAmount;
 
     @NotNull
     @Min(1)
