+++ src/main/java/ke/co/myfuture/Myfuture/Treasury/Loan/LoanScheduleItem.java
@@ -4,7 +4,6 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -13,8 +12,8 @@ import java.time.LocalDate;
 public class LoanScheduleItem {
     private int installmentNumber;
     private LocalDate dueDate;
-    private BigDecimal principal;
-    private BigDecimal interest;
-    private BigDecimal totalPayment;
-    private BigDecimal remainingBalance;
+    private Double principal;
+    private Double interest;
+    private Double totalPayment;
+    private Double remainingBalance;
 }
